preparing for arm relocs

+++ arm/arm_expr_types.h
@@ -16,9 +16,10 @@
 namespace kas::expression
 {
     // override default `kas` types in `expr_types.h`
-    // undefined addresses are 16-bits
-    template <> struct e_data_t <void> { using type = std::uint16_t;  };
-    template <> struct e_addr_t <void> { using type = std::uint16_t;  };
+    // arm7 addr is 32-bits
+    // arm8 addr is 64-bits
+    template <> struct e_data_t <void> { using type = std::uint32_t;  };
+    template <> struct e_addr_t <void> { using type = std::uint64_t;  };
     template <> struct err_msg_t<void> { using type = arm::error_msg; };
 }
